requirements.txt
asyncio
aiomysql
discord.py[voice]